main

+++ src/main/java/phongkham/DTO/PhieuNhapDTO.java
@@ -5,7 +5,7 @@ public class PhieuNhapDTO {
     private String MaNhaCungCap;
     private String NgayNhap;
     private String NguoiGiao;
-    private int TongTienNhap;
+    private float TongTienNhap;
 
     public PhieuNhapDTO(){
         MaPhieuNhap = "";
@@ -15,7 +15,7 @@ public class PhieuNhapDTO {
         TongTienNhap = 0;
     }
 
-    public PhieuNhapDTO(String MaPhieuNhap, String MaNhaCungCap, String NgayNhap, String NguoiGiao, int TongTienNhap){
+    public PhieuNhapDTO(String MaPhieuNhap, String MaNhaCungCap, String NgayNhap, String NguoiGiao, float TongTienNhap){
         this.MaPhieuNhap = MaPhieuNhap;
         this.MaNhaCungCap = MaNhaCungCap;
         this.NgayNhap = NgayNhap;
@@ -35,11 +35,11 @@ public class PhieuNhapDTO {
     public String getMaNhaCungCap() {return MaNhaCungCap;}
     public String getNgayNhap() {return NgayNhap;}
     public String getNguoiGiao() {return NguoiGiao;}
-    public int getTongTienNhap() {return TongTienNhap;}
+    public float getTongTienNhap() {return TongTienNhap;}
 
     public void setMaPhieuNhap(String MaPhieuNhap) {this.MaPhieuNhap = MaPhieuNhap;}
     public void setMaNhaCungCap(String MaNhaCungCap) {this.MaNhaCungCap = MaNhaCungCap;}
     public void setNgayNhap(String NgayNhap) {this.NgayNhap = NgayNhap;}
     public void setNguoiGiao(String NguoiGiao) {this.NguoiGiao = NguoiGiao;}
-    public void setTongTienNhap(int TongTienNhap) {this.TongTienNhap = TongTienNhap;}
+    public void setTongTienNhap(float TongTienNhap) {this.TongTienNhap = TongTienNhap;}
 }
