+++ src/main/java/phongkham/DTO/ThuocDTO.java
@@ -5,7 +5,7 @@ public class ThuocDTO {
     private String TenThuoc;
     private String HoatChat;
     private String DonViTinh;
-    private int DonGiaBan;
+    private float DonGiaBan;
     private int SoLuongTon;
 
     public ThuocDTO(){
@@ -17,7 +17,7 @@ public class ThuocDTO {
         SoLuongTon = 0;
     }
 
-    public ThuocDTO(String MaThuoc, String TenThuoc, String HoatChat, String DonViTinh, int DonGiaBan, int SoLuongTon){
+    public ThuocDTO(String MaThuoc, String TenThuoc, String HoatChat, String DonViTinh, float DonGiaBan, int SoLuongTon){
         this.MaThuoc = MaThuoc;
         this.TenThuoc = TenThuoc;
         this.HoatChat = HoatChat;
@@ -39,13 +39,13 @@ public class ThuocDTO {
     public String getTenThuoc() {return TenThuoc;}
     public String getHoatChat() {return HoatChat;}
     public String getDonViTinh() {return DonViTinh;}
-    public int getDonGiaBan() {return DonGiaBan;}
+    public float getDonGiaBan() {return DonGiaBan;}
     public int getSoLuongTon() {return SoLuongTon;}
 
     public void setMaThuoc(String MaThuoc) {this.MaThuoc = MaThuoc;}
     public void setTenThuoc(String TenThuoc) {this.TenThuoc = TenThuoc;}
     public void setHoatChat(String HoatChat) {this.HoatChat = HoatChat;}
     public void setDonViTinh(String DonViTinh) {this.DonViTinh = DonViTinh;}
-    public void setDonGiaBan(int DonGiaBan) {this.DonGiaBan = DonGiaBan;}
+    public void setDonGiaBan(float DonGiaBan) {this.DonGiaBan = DonGiaBan;}
     public void setSoLuongTon(int SoLuongTon) {this.SoLuongTon = SoLuongTon;}
 }
